Clean up race ethnicities - breakout components
